wip: clean up common+backend
