+++ halo2_backend/src/lib.rs
@@ -4,7 +4,7 @@ pub mod poly;
 
 // Internal re-exports
 pub use halo2_common::arithmetic;
-pub use halo2_common::circuit;
+// pub use halo2_common::circuit;
 pub use halo2_common::multicore;
 pub use halo2_common::transcript;
 pub use halo2_common::SerdeFormat;
