--- halo2_common/src/circuit/floor_planner.rs
@@ -1,6 +1,0 @@
-//! Implementations of common circuit floor planners.
-
-pub mod single_pass;
-
-pub mod v1;
-pub use v1::{V1Pass, V1};
