+++ halo2_common/src/lib.rs
@@ -7,7 +7,6 @@
 #![deny(unsafe_code)]
 
 pub mod arithmetic;
-pub mod circuit;
 pub use halo2curves;
 pub mod multicore;
 pub mod plonk;
