+++ halo2_frontend/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod circuit;
 pub mod dev;
+pub mod plonk;
