+++ halo2_frontend/src/plonk/circuit/compress_selectors.rs
